Remove static qualifiers from Server fields
Other than the instance of the Server class itself, all the static qualifiers have been removed. This should make it possible to have multiple instances of the server, and also it's a step on the way to make the server more testable, and runnable as a thread instead of a separate process in standalone mode.