Logging library + refactor (#36)
Tinylog library added to the project. Nearly all the occurrences of System.(err|out).println were replaced with Logger calls, similarly existing logging class was readapted. Some of the log levels likely do not reflect severity correctly - that's something to fix as we go.
Threads got names now, as Tinylog displays thread name in the message by default, which seems to be a good idea.