Surrendering tutorial is now skipping it (#38)
* Surrendering tutorial is now skipping it

A bit of a hack, but for now the easiest way to implement skipping the
tutorial. The hack is mostly regarding introducing more calls to the
Client's static methods - conceptually it's fine in the sense of
surrendering not being very meaningful in a tutorial game.

* Client/Server interfacing change - protocol update

While it's not an incompatibility of the communication protocol, results of surrendering in tutorial might differ from expected, so a protocol version bump makes it clean and clear.